+++ akeneo-design-system/src/components/Modal/Modal.tsx
@@ -70,6 +70,19 @@ const Title = styled.div`
   margin-bottom: 10px;
 `;
 
+const BottomButtons = styled.div`
+  display: flex;
+  gap: 10px;
+  margin-top: 20px;
+`;
+
+const TopRightButtons = styled(BottomButtons)`
+  position: absolute;
+  top: 40px;
+  right: 40px;
+  margin: 0;
+`;
+
 type ModalProps = Override<
   React.HTMLAttributes<HTMLDivElement>,
   {
@@ -103,7 +116,10 @@ type ModalProps = Override<
 /**
  * The Modal Component is used to display a secondary window over the content.
  */
-const Modal = ({isOpen, onClose, illustration, closeTitle, children, ...rest}: ModalProps) => {
+const Modal: React.FC<ModalProps> & {
+  BottomButtons: typeof BottomButtons;
+  TopRightButtons: typeof TopRightButtons;
+} = ({isOpen, onClose, illustration, closeTitle, children, ...rest}: ModalProps) => {
   useShortcut(Key.Escape, onClose);
 
   const portalNode = document.createElement('div');
@@ -136,17 +152,7 @@ const Modal = ({isOpen, onClose, illustration, closeTitle, children, ...rest}: M
   );
 };
 
-Modal.BottomButtons = styled.div`
-  display: flex;
-  gap: 10px;
-  margin-top: 20px;
-`;
-
-Modal.TopRightButtons = styled(Modal.BottomButtons)`
-  position: absolute;
-  top: 40px;
-  right: 40px;
-  margin: 0;
-`;
+Modal.BottomButtons = BottomButtons;
+Modal.TopRightButtons = TopRightButtons;
 
 export {Modal, SectionTitle, Title};